% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testGeneticNhoods.R
\name{dot-expandRandomDesign}
\alias{dot-expandRandomDesign}
\alias{.expandRandomDesign}
\title{Expand a random effect design into indicator columns}
\usage{
.expandRandomDesign(z)
}
\arguments{
\item{z}{A \code{matrix} or \code{data.frame} of grouping variables, one
column per random effect.}
}
\value{
A \code{matrix} of indicator columns with a \code{u_indices}
attribute giving the 1-based column indices belonging to each random effect.
}
\description{
Internal helper turning a matrix of random effect grouping variables into a
binary indicator design matrix, recording which columns belong to which
random effect.
}
\examples{
zz <- matrix(c(1, 1, 2, 2), ncol=1, dimnames=list(NULL, "batch"))
miloR:::.expandRandomDesign(zz)

}
\author{
Mike Morgan
}
