% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fitPLGlmm}
\alias{fitPLGlmm}
\title{GLMM parameter estimation using pseudo-likelihood}
\usage{
fitPLGlmm(
  Z,
  X,
  muvec,
  offsets,
  curr_beta,
  curr_theta,
  curr_u,
  curr_sigma,
  curr_G,
  y,
  u_indices,
  theta_conv,
  rlevels,
  curr_disp,
  REML,
  maxit
)
}
\arguments{
\item{Z}{mat - sparse matrix that maps random effect variable levels to
observations}

\item{X}{mat - sparse matrix that maps fixed effect variables to
observations}

\item{muvec}{vec vector of estimated phenotype means}

\item{offsets}{vec vector of model offsets}

\item{curr_beta}{vec vector of initial beta estimates}

\item{curr_theta}{vec vector of initial parameter estimates}

\item{curr_u}{vec of initial u estimates}

\item{curr_sigma}{vec of initial sigma estimates}

\item{curr_G}{mat c X c matrix of variance components}

\item{y}{vec of observed counts}

\item{u_indices}{List a List, each element contains the indices of Z relevant
to each RE and all its levels}

\item{theta_conv}{double Convergence tolerance for paramter estimates}

\item{rlevels}{List containing mapping of RE variables to individual
levels}

\item{curr_disp}{double Dispersion parameter estimate}

\item{REML}{bool - use REML for variance component estimation}

\item{maxit}{int maximum number of iterations if theta_conv is FALSE}
}
\description{
Iteratively estimate GLMM fixed and random effect parameters, and variance
component parameters using Fisher scoring based on the Pseudo-likelihood
approximation to a Normal loglihood.
}
