% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{invertPseudoVar}
\alias{invertPseudoVar}
\title{Compute the inverse of a structured covariance matrix}
\usage{
invertPseudoVar(A, B, Z)
}
\arguments{
\item{A}{SparseMatrix - a nxn matrix of the GLMM covariance D^-1*V*D^-1}

\item{B}{SparseMatrix - a cxc matrix of variance components}

\item{Z}{SparseMatrix - a nxc design matrix that maps REs to samples}
}
\description{
Using Henderson's adjusted Woodbury formula for a singular B matrix,
compute the inverse of the pseudocovariance matrix ZGZ' + W as
(A + UBU^T)^-1 = A^-1 - A^-1UB[I + U^TA^-1UB]^-1U^TA^-1
}
