% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miloR-package.R
\name{miloR-package}
\alias{miloR-package}
\title{The miloR package}
\description{
The \pkg{miloR} package provides modular functions to perform differential
abundance testing on replicated single-cell experiments. For details please
see the vignettes \code{vignette("milo_demo", package="miloR")} and
\code{vignette("milo_gastrulation", package="miloR")}.
}
\author{
Mike Morgan & Emma Dann
}
