% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miloR.R
\docType{package}
\name{miloR}
\alias{miloR}
\title{miloR}
\description{
Milo performs single-cell differential abundance testing. Cell states are modelled
as representative neighbourhoods on a nearest neighbour graph. Hypothesis testing is performed using a
negative bionomial generalized linear model.
}
