% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{multiP}
\alias{multiP}
\title{Compute product of each pseudovariance partial derivatives with the inverse
pseudovariance matrix}
\usage{
multiP(partials, psvar_in)
}
\arguments{
\item{List}{partials - list containing matrices of partial derivatives of the pseudovariance
for each variance component}

\item{SparseMatrix}{psvar_in - inverse of the pseudovariance matrix}
}
\description{
For each variance component, we compute the matrix multiplication of the
relevant partial derivative of dV_start/dSigm with the pseudovariance matrix
}
