% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotNhoods.R
\name{plotMiloReducedDim}
\alias{plotMiloReducedDim}
\title{Plot Milo test results on reduced dimensiona}
\usage{
plotMiloReducedDim(
  x,
  milo_results,
  nhood_reduced_dims = "UMAP",
  filter_alpha = NULL,
  split_by = NULL,
  pt_size = 1.5,
  components = c(1, 2)
)
}
\arguments{
\item{x}{A \code{\linkS4class{Milo}} object}

\item{milo_results}{A `data.frame` containing the results of differential nhood abundance testing (output of \code{testNhoods})
--> this will need to be changed/removed when output of testNhoods changes}

\item{nhood_reduced_dims}{a character indicating the name of the \code{reducedDim} slot in the
\code{\linkS4class{Milo}} object to use as (default: 'UMAP').}

\item{filter_alpha}{the spatialFDR cutoff used as a significance threshold. If not \code{NULL} the logFC will be plotted only for
significantly DA nhoods (default: NULL)}

\item{split_by}{A character indicating the \code{colData} column in \code{x} to use for faceting
e.g. useful to visualize results by cell type}

\item{pt_size}{size of scatterplot points (default: 1.5)}

\item{components}{vector of reduced dimensions components to plot (default: c(1,2))}
}
\value{
a \code{ggplot-class} object
}
\description{
Visualize log-FC estimated with differential nhood abundance testing
on embedding of original single-cell dataset.
}
\examples{
NULL

}
\author{
Emma Dann
}
