% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotNhoods.R
\name{plotNhoodExpressionDA}
\alias{plotNhoodExpressionDA}
\title{Visualize gene expression in neighbourhoods}
\usage{
plotNhoodExpressionDA(
  x,
  da.res,
  features,
  alpha = 0.1,
  subset.nhoods = NULL,
  cluster_features = FALSE,
  assay = "logcounts",
  scale_to_1 = FALSE
)
}
\arguments{
\item{x}{A \code{\linkS4class{Milo}} object}

\item{da.res}{a data.frame of DA testing results}

\item{features}{a character vector of features to plot (they must be in rownames(x))}

\item{alpha}{significance level for Spatial FDR (default: 0.1)}

\item{subset.nhoods}{A logical, integer or character vector indicating a subset of nhoods to show in plot 
(default: NULL, no subsetting)}

\item{cluster_features}{logical indicating whether features should be clustered with hierarchical clustering.
If FALSE then the order in \code{features} is maintained (default: FALSE)}

\item{assay}{A character scalar that describes the assay slot to use for calculating neighbourhood expression. 
(default: logcounts)
Of note: neighbourhood expression will be computed only if the requested features are not in the \code{nhoodExpression} slot
of the milo object. If you wish to plot average neighbourhood expression from a different assay, you should run 
\code{calcNhoodExpression(x)} with the desired assay.}
}
\value{
a \code{\linkS4class{ggplot}} object
}
\description{
Plots the average gene expression in neighbourhoods, sorted by DA fold-change
}
\examples{
NULL

}
\author{
Emma Dann
}
