% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotNhoods.R
\name{plotNhoodGraphDA}
\alias{plotNhoodGraphDA}
\alias{plotNhoodGroups}
\title{Plot Milo results on graph of neighbourhood}
\usage{
plotNhoodGraphDA(x, milo_res, alpha = 0.05, res_column = "logFC", ...)

plotNhoodGroups(x, milo_res, show_groups = NULL, ...)
}
\arguments{
\item{x}{A \code{\linkS4class{Milo}} object}

\item{milo_res}{a data.frame of milo results containing the \code{nhoodGroup} column}

\item{alpha}{significance level for Spatial FDR (default: 0.05)}

\item{res_column}{which column of \code{milo_res} object to use for color (default: logFC)}

\item{...}{arguments to pass to \code{plotNhoodGraph}}

\item{show_groups}{a character vector indicating which groups to plot
all other neighbourhoods will be gray}
}
\value{
a \code{ggplot} object

a \code{ggplot} object
}
\description{
Visualize log-FC estimated with differential nhood abundance testing
on embedding of original single-cell dataset.

Visualize grouping of neighbourhoods obtained with \code{groupNhoods}
}
\examples{
NULL

NULL

}
\author{
Emma Dann
}
