% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotNhoods.R
\name{plotNhoodGroups}
\alias{plotNhoodGroups}
\title{Plot graph of neighbourhoods coloring by nhoodGroups}
\usage{
plotNhoodGroups(x, milo_res, show_groups = NULL, ...)
}
\arguments{
\item{x}{A \code{\linkS4class{Milo}} object}

\item{milo_res}{a data.frame of milo results containing the \code{nhoodGroup} column}

\item{show_groups}{a character vector indicating which groups to plot
all other neighbourhoods will be gray}

\item{...}{arguments to pass to \code{plotNhoodGraph}}
}
\value{
a \code{ggplot} object
}
\description{
Visualize grouping of neighbourhoods obtained with \code{groupNhoods}
}
\examples{
NULL

}
\author{
Emma Dann
}
