% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projectNhoodExpression.R
\name{projectNhoodExpression}
\alias{projectNhoodExpression}
\title{Project mean neighbourhood expression profiles on embedding of single-cells}
\usage{
projectNhoodExpression(
  x,
  d = 30,
  reduced_dims = "PCA",
  scale = TRUE,
  center = TRUE
)
}
\arguments{
\item{x}{A \code{\linkS4class{Milo}} object with a non-empty \code{nhoods}
slot.}

\item{d}{The number of reduced dimensions to use}

\item{reduced_dims}{a character indicating the name of the \code{reducedDim} slot in the
\code{\linkS4class{Milo}} object to use  for projection (default: 'PCA'). \code{reducedDim(x,reduced_dim)}
needs to have the attribute \code{rotation} storing the gene X dimensions loading matrix to use for projection.
See details for more info.}

\item{scale}{logical indicating whether nhoods expression profiles should be scaled before projection.
Data should be scaled if the embedding was performed on scaled single-cell data.
(Default: TRUE)}

\item{center}{logical indicating whether nhoods expression profiles should be centered before projection
Data should be centered if the embedding was performed on scaled single-cell data.
(Default: TRUE)}
}
\value{
A \code{\linkS4class{matrix}} object of samples X reduced dimensions where samples are both
the nhoods and the single-cells.
(This will need to become a slot in Milo object)
}
\description{
This is for visualization of single-cells and nhoods on the same embedding
}
\details{
This function projects nhoods in the same reduced dimensions used for the full single-cell dataset.
First it calculates the mean expression profile for all cells in a nhood. Then the gene X nhood matrix is
multiplied by the matrix of loadings of genes on reduced dimensions. The loading matrix should be stored as an attribute \code{rotation}
in \code{reducedDim(x, reduced_dims)}. This can be added by running \code{runPCA(x)} from the \code{scater} package or, using the
package \code{irlba} as in examples. The output of this function is used for visualization of test results.
}
\examples{

require(SingleCellExperiment)
ux <- matrix(rpois(12000, 5), ncol=200)
vx <- log2(ux + 1)
pca <- prcomp(t(vx))
rotation <- pca$rotation
milo <- Milo(SingleCellExperiment(assays=list(logcounts=vx), reducedDims=SimpleList(PCA=pca$x)))
milo <- buildGraph(milo, k=10, d=10, transposed=TRUE)
milo <- makeNhoods(milo, d=10)
attr(reducedDim(milo, "PCA"), "rotation") <- rotation
milo <- projectNhoodExpression(milo, d=10)

}
\author{
Emma Dann
}
