% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{pseudovarPartial}
\alias{pseudovarPartial}
\title{Compute pseudovariance partial derivatives}
\usage{
pseudovarPartial(x, rlevels, cnames)
}
\arguments{
\item{Matrix}{x - the fully expanded Z matrix that maps observations to
random effect variables}

\item{List}{dimnames - a list of the matrix `x` dimension names.}
}
\description{
Compute the partial derivatives of the pseudovariance as t(Z[, i]) %*% Z[, i]
for the ith variance component
}
