% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmm.R, R/glmm_new.R
\name{runGLMM}
\alias{runGLMM}
\title{Perform differential abundance testing using a NB-generalised linear mixed model}
\usage{
runGLMM(
  X,
  Z,
  y,
  init.theta = NULL,
  crossed = FALSE,
  random.levels = NULL,
  REML = FALSE,
  glmm.control = list(theta.tol = 1e-06, max.iter = 100),
  dispersion = 0.5
)

runGLMM(
  X,
  Z,
  y,
  init.theta = NULL,
  crossed = FALSE,
  random.levels = NULL,
  REML = FALSE,
  glmm.control = list(theta.tol = 1e-06, max.iter = 100),
  dispersion = 0.5
)
}
\arguments{
\item{x}{A \code{\linkS4class{Milo}} object with a non-empty
\code{nhoodCounts} slot.}

\item{error.model}{A string vector dictating the type of error model to use for the LMM - either
'normal' (default) or 'negbinom'. The former should only be used for approximately normally distributed
input variables, and the latter for overdispersed counts.}
}
\description{
This function will perform DA testing on all nhoods using a negative binomial generalised linear mixed model

This function will perform DA testing on all nhoods using a negative binomial generalised linear mixed model
}
