% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_discrete.R
\docType{data}
\name{sim_discrete}
\alias{sim_discrete}
\title{sim_discrete}
\format{
A list containing a \code{\linkS4class{Milo}} object in the "mylo" slot,
and a \code{data.frame} containing experimental meta-data in the "meta" slot.
}
\usage{
data(sim_discrete)
}
\description{
Simulated discrete groups data
}
\details{
Data are simulated single-cells in 4 distinct groups of cells. Cells in each
group are assigned to 1 of 2 conditions: \emph{A} or \emph{B}. Specifically,
the cells in block 1 are highly abundant in the \emph{A} condition, whilst
cells in block 4 are most abundant in condition \emph{B}.
}
\examples{
NULL

}
\keyword{datasets}
