% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_trajectory.R
\docType{data}
\name{sim_trajectory}
\alias{sim_trajectory}
\title{Simulated linear trajectory data}
\format{
A list containing a \code{\linkS4class{Milo}} object in the "mylo" slot,
and a \code{data.frame} containing experimental meta-data in the "meta" slot.
}
\usage{
data(sim_trajectory)
}
\description{
Data are simulated single-cells along a single linear trajectory. Cells are
simulated from 5 groups, and assigned to 1 of 2 conditions; \emph{A} or \emph{B}.
Data were generated using in the \code{simulate_linear_trajectory}
function in the \code{dyntoy} package.
}
\examples{
NULL

}
\references{
https://github.com/dynverse/dyntoy
}
\keyword{datasets}
