% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildNhoodGraph.R
\name{buildNhoodGraph}
\alias{buildNhoodGraph}
\title{Build an abstracted graph of neighbourhoods for visualization}
\usage{
buildNhoodGraph(x)
}
\arguments{
\item{x}{A \code{\linkS4class{Milo}} object with a non-empty \code{nhoods}
slot.}
}
\value{
A \code{\linkS4class{Milo}} object with filled
}
\description{
Build an abstracted graph of neighbourhoods for visualization
}
\details{
This constructs a graph where nodes represent neighbourhoods and edges represent the number of overlapping
cells between two neighbourhoods.
}
\examples{

NULL

}
\author{
Emma Dann
}
