% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmm.R
\name{fitGLMM}
\alias{fitGLMM}
\title{Perform differential abundance testing using a NB-generalised linear mixed model}
\usage{
fitGLMM(
  X,
  Z,
  y,
  offsets,
  init.theta = NULL,
  Kin = NULL,
  random.levels = NULL,
  REML = FALSE,
  glmm.control = list(theta.tol = 1e-06, max.iter = 100, init.sigma = NULL, init.beta =
    NULL, init.u = NULL),
  dispersion = 0.5,
  geno.only = FALSE,
  solver = NULL
)
}
\arguments{
\item{X}{A matrix containing the fixed effects of the model.}

\item{Z}{A matrix containing the random effects of the model.}

\item{y}{A matrix containing the observed phenotype over each neighborhood.}

\item{offsets}{A vector containing the (log) offsets to apply normalisation for different numbers of cells across samples.}

\item{init.theta}{A column vector (m X 1 matrix) of initial estimates of fixed and random effect coefficients}

\item{Kin}{A n x n covariance matrix to explicitly model variation between observations}

\item{random.levels}{A list describing the random effects of the model, and for each, the different unique levels.}

\item{REML}{A logical value denoting whether REML (Restricted Maximum Likelihood) should be run. Default is TRUE.}

\item{glmm.control}{A list containing parameter values specifying the theta tolerance of the model, the maximum number of iterations to be run,
and initial parameter values for the fixed (init.beta) and random effects (init.u)}

\item{dispersion}{A scalar value for the dispersion of the negative binomial.}

\item{geno.only}{A logical value that flags the model to use either just the \code{matrix} `Kin` or the supplied random effects.}

\item{solver}{a character value that determines which optmisation algorithm is used for the variance components. Must be either
HE (Haseman-Elston regression) or Fisher (Fisher scoring).}
}
\description{
This function will perform DA testing per-nhood using a negative binomial generalised linear mixed model
}
\details{
This function runs a negative binomial generalised linear mixed effects model. If mixed effects are detected in testNhoods,
this function is run to solve the model.
}
