% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmm.R
\name{fitGLMM}
\alias{fitGLMM}
\title{Wrapper function for fitPLGlmm}
\usage{
fitGLMM(
  X,
  full.Z,
  y,
  offsets,
  init.theta = NULL,
  Kin = NULL,
  random.levels = NULL,
  REML = FALSE,
  glmm.control = list(theta.tol = 1e-06, max.iter = 100),
  dispersion = 0.5
)
}
\description{
Wrapper function for fitPLGlmm
}
