% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{matrixORMatrix-class}
\alias{matrixORMatrix-class}
\title{The Milo container class}
\description{
The Milo container class
}
\section{Slots}{

\describe{
\item{\code{graph}}{An igraph object that represents the kNN graph}

\item{\code{nhoods}}{A CxN binary sparse matrix mapping cells to the neighbourhoods they belong to}

\item{\code{nhoodDistances}}{An list of PxN sparse matrices of Euclidean distances
between vertices in each neighbourhood, one matrix per neighbourhood}

\item{\code{nhoodCounts}}{An NxM sparse matrix of cells counts in each neighourhood
across M samples}

\item{\code{nhoodIndex}}{A list of the index vertices for each neighbourhood}

\item{\code{nhoodExpression}}{An GxN matrix of genes X neighbourhoods containing
average gene expression levels across cells in each neighbourhood}

\item{\code{nhoodReducedDim}}{a list of reduced dimensional representations of
neighbourhoods, including projections into lower dimension space}

\item{\code{nhoodGraph}}{an igraph object that represents the graph of neighbourhoods}

\item{\code{.k}}{A hidden slot that stores the value of k used for graph building}
}}

