% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotNhoods.R
\name{plotNhoodGraphDA}
\alias{plotNhoodGraphDA}
\title{Plot Milo results on graph of neighbourhood}
\usage{
plotNhoodGraphDA(x, milo_res, alpha = 0.05, ...)
}
\arguments{
\item{x}{A \code{\linkS4class{Milo}} object}

\item{milo_res}{a data.frame of milo results}

\item{alpha}{significance level for Spatial FDR (default: 0.05)}

\item{...}{arguments to pass to \code{plotNhoodGraph}}
}
\value{
a \code{ggplot} object
}
\description{
Visualize log-FC estimated with differential nhood abundance testing
on embedding of original single-cell dataset.
}
\examples{
NULL

}
\author{
Emma Dann
}
