% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{pseudovarPartial}
\alias{pseudovarPartial}
\title{Compute pseudovariance partial derivatives}
\usage{
pseudovarPartial(x, rlevels, cnames)
}
\arguments{
\item{matrix}{x - the fully expanded Z matrix that maps observations to
random effect variables}

\item{list}{rlevels - a list that maps the random effect variable to the
individual levels}

\item{vec}{cnames - a StringVector of column names from fully expanded Z matrix}
}
\description{
Compute the partial derivatives of the pseudovariance as t(Z[, i]) %*% Z[, i]
for the ith variance component
}
