% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmm.R
\name{runGLMM}
\alias{runGLMM}
\title{Perform differential abundance testing using a NB-generalised linear mixed model}
\usage{
runGLMM(
  X,
  Z,
  y,
  random.levels = NULL,
  REML = TRUE,
  glmm.control = list(theta.tol = 1e-06, max.iter = 100),
  dispersion = NULL
)
}
\arguments{
\item{X}{A matrix containing the fixed effects of the model.}

\item{Z}{A matrix containing the random effects of the model.}

\item{y}{A matrix containing the observed phenotype over each neighborhood.}

\item{random.levels}{A list describing the random effects of the model, and for each, the different unique levels.}

\item{REML}{A logical value denoting whether REML (Restricted Maximum Likelihood) should be run. Default is TRUE.}

\item{glmm.control}{A list containing parameter values specifying the theta tolerance of the model and the maximum number of iterations to be run.}

\item{dispersion}{A scalar value for the dispersion of the negative binomial.}
}
\description{
This function will perform DA testing on all nhoods using a negative binomial generalised linear mixed model
}
\details{
This function runs a negative binomial generalised linear mixed effects model. If mixed effects are detected in testNhoods,
this function is run to solve the model.
}
