% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmm.R
\name{runGLMM}
\alias{runGLMM}
\title{Perform differential abundance testing using a NB-generalised linear mixed model}
\usage{
runGLMM(
  X,
  Z,
  y,
  random.levels = NULL,
  REML = TRUE,
  glmm.control = list(theta.tol = 1e-06, max.iter = 50),
  dispersion = NULL
)
}
\arguments{
\item{X}{fixed effects model matrix}

\item{Z}{random effects model matrix}

\item{y}{observed phenotype}

\item{random.levels}{levels of random effects}

\item{REML}{TRUE or FALSE}

\item{glmm.control}{list containing parameter values}

\item{dispersion}{r dispersion}
}
\description{
This function will perform DA testing on all nhoods using a negative binomial generalised linear mixed model
}
